// generated/src/app/api-models/determinant-values.model.ts
/* Copyright (c) 2020 . All Rights Reserved. */

export class DeterminantValues {
  keyword: string;
  seqRuleId: number;
  searchSequence: number;
  determinantSequence: number;
  operator: string;
  valueFrom: string;
  valueThru: string;
  secDeterminantTable: string;
  secDeterminantColumn: string;
  secDeterminantDatatype: string;
  securityCode: string;
  insertDatetime: Date;
  insertUser: string;
  insertProcess: string;
  updateDatetime: Date;
  updateUser: string;
  updateProcess: string;
}